# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Uffizi Ticket App API
  description: |
    API for managing Uffizi Gallery tour bookings.

    This API syncs bookings from Bokun, tracks Uffizi ticket purchases,
    and manages multi-channel messaging (WhatsApp, SMS, Email) for sending
    tickets to customers.

    ## Authentication
    All endpoints except `/health`, `/health/detailed`, `/login`, and webhooks
    require Bearer token authentication via Laravel Sanctum.

    Obtain a token via `POST /api/login` and include it in subsequent requests:
    ```
    Authorization: Bearer {token}
    ```

    ## Rate Limiting
    - Login: 5 requests per minute
    - General API: 60 requests per minute
    - Sync/Import: 10 requests per minute

    ## Timezone
    All dates are in UTC. The frontend uses Europe/Rome for display.
  version: 2.0.0
  contact:
    name: API Support
    email: support@florencewithlocals.com
  license:
    name: Proprietary
    url: https://uffizi.deetech.cc
servers:
  - url: https://uffizi.deetech.cc/api
    description: Production server
  - url: http://localhost:8000/api
    description: Local development server
...
paths:
...
  /bookings/{id}/send-ticket:
    post:
      tags:
        - Messages
      summary: Send ticket to customer
      description: |
        Send ticket via the detected channel (WhatsApp, SMS, or Email).
        Requires ticket reference and attached PDF.
      operationId: sendTicket
      parameters:
        - $ref: '#/components/parameters/BookingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendTicketRequest'
      responses:
        '200':
          description: Ticket sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendTicketResponse'
        '400':
          description: Cannot send ticket (missing requirements)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{id}/detect-channel:
    get:
      tags:
        - Messages
      summary: Detect messaging channel
      description: Check which messaging channel is available for this booking.
      operationId: detectChannel
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Channel detection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelDetectionResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /templates/languages:
    get:
      tags:
        - Templates
      summary: Get supported languages
      description: Get list of supported languages with metadata.
      operationId: getLanguages
      responses:
        '200':
          description: List of languages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                      example: en
                    name:
                      type: string
                      example: English
                    flag:
                      type: string
                      example: "\U0001F1EC\U0001F1E7"
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Laravel Sanctum Bearer Token
  parameters:
    BookingId:
      name: id
      in: path
      required: true
      description: Booking ID
      schema:
        type: integer
  responses:
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthenticated."
              error:
                type: string
                example: "Please login to access this resource."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Resource not found"
  schemas:
...
    SendTicketRequest:
      type: object
      required:
        - template_id
      properties:
        template_id:
          type: integer
          description: ID of the message template to use
        reference_number:
          type: string
          description: Uffizi ticket reference (if not already set)
        attachment_ids:
          type: array
          items:
            type: integer
          description: IDs of PDF attachments to include
        custom_message:
          type: string
          description: Custom message content (overrides template)
    SendTicketResponse:
      type: object
      properties:
        success:
          type: boolean
        message_id:
          type: integer
        channel:
          type: string
          enum: [whatsapp, sms, email]
        status:
          type: string
        external_id:
          type: string
          description: Twilio SID or mail ID
    ChannelDetectionResponse:
      type: object
      properties:
        has_phone:
          type: boolean
        has_email:
          type: boolean
        whatsapp_available:
          type: boolean
        recommended_channel:
          type: string
          enum: [whatsapp, email_sms, email_only]
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
...
security:
  - bearerAuth: []